Time: 0 ms (100%), Space: 41.7 MB (67.48%) - LeetHub

+++ 0027-remove-element/0027-remove-element.java
@@ -1,7 +1,7 @@
 //5
 class Solution {
     public int removeElement(int[] nums, int val){
-        short j = 0;
+        int j = 0;
         int l = nums.length;
         for(int i=0; i<l; i++){
             if(nums[i] != val){
